+++ epde/operators/moeadd_specific.py
@@ -1,0 +1,307 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+"""
+Created on Fri Jul 29 19:08:51 2022
+
+@author: maslyaev
+"""
+import copy
+import numpy as np
+from functools import reduce
+
+from epde.moeadd.moeadd_stc import Constraint
+from epde.operators.template import CompoundOperator
+
+
+def penalty_based_intersection(sol_obj, weight, ideal_obj, penalty_factor = 1.) -> float:
+    '''
+    Calculation of the penalty pased intersection, that is minimized for the solutions inside the 
+    domain, specified by **weight** vector. The calculations are held, according to the following formulas:
+        
+    .. math:: g^{pbi}(\mathbf{x}|\mathbf{w}, \mathbf{z^{*}}) = d_1 + \Theta d_2 \longrightarrow min
+        
+    subject to :math:`\mathbf{x} \in \Omega`
+
+    where: 
+        
+    .. math::        
+        d_1 = ||(\mathbf{f}(\mathbf{x}) - \mathbf{z^{*}})^{t}\mathbf{w}|| (||\mathbf{w}||)^{-1}
+
+        d_2 = || \mathbf{f}(\mathbf{x}) - (\mathbf(z^{*}) + d_1 \mathbf{w} (||\mathbf{w}||)^{-1})||
+
+    Arguments:
+    ----------
+    
+    sol_obj : object of subclass of ``src.moeadd.moeadd_stc.moeadd_solution``
+        The solution, for which the penalty based intersection is calculated. In the equations above,
+        it denotes :math:`\mathbf{x}`, with the :math:`\mathbf{F}(\mathbf{x})` representing the
+        objective function values.
+    
+    weight : np.array
+        Values of the weight vector, specific to the domain, in which the solution is located.
+        Represents the :math:`\mathbf{w}` in the equations above.
+    
+    ideal_obj : `np.array`
+        The value of best achievable objective functions values; denoted as 
+        :math:`\mathbf{z^{*}} = (z^{*}_1, z^{*}_2, \; ... \;, z^{*}_m)`.
+    
+    penalty_factor : float, optional, default 1.
+        The penalty parameter, represents :math:`\Theta` in the equations.
+    
+    '''
+    d_1 = np.dot((sol_obj.obj_fun - ideal_obj), weight) / np.linalg.norm(weight)
+    d_2 = np.linalg.norm(sol_obj.obj_fun - (ideal_obj + d_1 * weight/np.linalg.norm(weight)))
+    return d_1 + penalty_factor * d_2
+
+
+def population_to_sectors(population, weights):
+    '''
+    
+    The distribution of the solutions into the domains, defined by weights vectors.
+    
+    Parameters:
+    -----------
+    
+    population : list
+        List, containing the candidate solutions for the evolutionary algorithm. Elements shall
+        belong to the case-specific subclass of ``src.moeadd.moeadd_stc.moeadd_solution``.
+        
+    weights : np.ndarray
+        Numpy ndarray of weight vectors; first dimension - weight index, second dimension - 
+        weight value in the objective function space.
+        
+    Returns:
+    ---------
+    
+    population_divided : list
+        List of candidate solutions, belonging to the weight domain. The outer index of the list - 
+        the weight vector index, inner - the index of a particular candidate solution inside the domain.
+
+        
+    '''
+    solution_selection = lambda weight_idx: [solution for solution in population if solution.get_domain(weights) == weight_idx]
+    return list(map(solution_selection, np.arange(len(weights))))    
+
+
+def locate_pareto_worst(levels, weights, best_obj, penalty_factor = 1.):
+    '''
+    
+    Function, dedicated to the selection of the worst solution on the Pareto levels.
+    
+    Arguments:
+    ----------
+    
+    levels : pareto_levels obj
+        The levels, on which the worst candidate solution is detected.
+    
+    weights : np.ndarray
+        The weight vectors of the moeadd optimizer.
+        
+    best_obj : np.array
+        Best achievable values of the objective functions.
+    
+    penalty_factor : float, optional, default 1.
+        The penalty parameter, used during penalty based intersection value calculation.        
+    
+    '''
+    domain_solutions = population_to_sectors(levels.population, weights)
+    most_crowded_count = max([len(domain) for domain in domain_solutions]); crowded_domains = [domain_idx for domain_idx in np.arange(len(weights)) if 
+                                                                           len(domain_solutions[domain_idx]) == most_crowded_count]
+    if len(crowded_domains) == 1:
+        most_crowded_domain = crowded_domains[0]
+    else:
+        PBI = lambda domain_idx: sum([penalty_based_intersection(sol_obj, weights[domain_idx], best_obj, penalty_factor) for sol_obj in domain_solutions[domain_idx]])
+        PBIS = np.fromiter(map(PBI, crowded_domains), dtype = float)
+        most_crowded_domain = crowded_domains[np.argmax(PBIS)]
+        
+    worst_NDL_section = []
+    domain_solution_NDL_idxs = np.empty(most_crowded_count)
+    for solution_idx, solution in enumerate(domain_solutions[most_crowded_domain]):
+        domain_solution_NDL_idxs[solution_idx] = [level_idx for level_idx in np.arange(len(levels.levels)) 
+                                                    if any([solution == level_solution for level_solution in levels.levels[level_idx]])][0]
+        
+    max_level = np.max(domain_solution_NDL_idxs)
+    worst_NDL_section = [domain_solutions[most_crowded_domain][sol_idx] for sol_idx in np.arange(len(domain_solutions[most_crowded_domain])) 
+                        if domain_solution_NDL_idxs[sol_idx] == max_level]
+    PBIS = np.fromiter(map(lambda solution: penalty_based_intersection(solution, weights[most_crowded_domain], best_obj, penalty_factor), worst_NDL_section), dtype = float)
+    return worst_NDL_section[np.argmax(PBIS)]
+
+
+class SimpleNeighborSelector(CompoundOperator):
+    def apply(self, neighbors):
+        '''
+            Simple selector of neighboring weight vectors: takes n-closest (*n = number_of_neighbors*)ones to the 
+            processed one. Defined to be used inside the moeadd algorithm.
+        
+            Arguments:
+            ----------
+            
+            sorted_neighbors : list
+                proximity list of neighboring vectors, ranged in the ascending order of the angles between vectors.
+                
+            number_of_neighbors : int
+                numbers of vectors to be considered as the adjacent ones
+                
+            Returns:
+            ---------
+            
+            sorted_neighbors[:number_of_neighbors] : list
+                self evident slice of proximity list
+        '''
+        return neighbors[:self.params['number_of_neighbors']]
+
+
+class PopulationUpdater(CompoundOperator):
+    def apply(self, objective, best_obj, weights):
+        '''
+        Update population to get the pareto-nondomiated levels with the worst element removed. 
+        Here, "worst" means the solution with highest PBI value (penalty-based boundary intersection)
+        '''         
+        # objective = {offspring, pareto_levels}        
+        
+        objective[1].update(objective[0])  #levels_updated = ndl_update(offspring, levels)
+        if len(objective[1].levels) == 1:
+            worst_solution = self.suboperators['locate_pareto_worst'].apply(objective[1], weights, 
+                                                                            best_obj, self.params['PBI_penalty'])
+        else:
+            if objective[1].levels[len(objective[1].levels) - 1] == 1:
+                domain_solutions = self.suboperators['population_to_sectors'].apply(objective[1].population, weights)
+                reference_solution = objective[1].levels[len(objective[1].levels) - 1][0]
+                reference_solution_domain = [idx for idx in np.arange(domain_solutions) if reference_solution in domain_solutions[idx]]
+                if len(domain_solutions[reference_solution_domain] == 1):
+                    worst_solution = self.suboperators['locate_pareto_worst'].apply(objective[1].levels, weights,
+                                                                                    best_obj, self.params['PBI_penalty'])                            
+                else:
+                    worst_solution = reference_solution
+            else:
+                last_level_by_domains = self.suboperators['population_to_sectors'].apply(objective[1].levels[len(objective[1].levels)-1], weights)
+                most_crowded_count = np.max([len(domain) for domain in last_level_by_domains]); 
+                crowded_domains = [domain_idx for domain_idx in np.arange(len(weights)) 
+                                   if len(last_level_by_domains[domain_idx]) == most_crowded_count]
+        
+                if len(crowded_domains) == 1:
+                    most_crowded_domain = crowded_domains[0]
+                else:
+                    PBI = lambda domain_idx: np.sum([self.suboperators.apply['penalty_based_intersection'].apply(sol_obj, weights[domain_idx],
+                                                                                                                 best_obj, self.params['PBI_penalty'])
+                                                     for sol_obj in last_level_by_domains[domain_idx]])
+                    PBIS = np.fromiter(map(PBI, crowded_domains), dtype = float)
+                    most_crowded_domain = crowded_domains[np.argmax(PBIS)]
+                    
+                if len(last_level_by_domains[most_crowded_domain]) == 1:
+                    worst_solution = self.suboperators['locate_pareto_worst'].apply(self.pareto_levels, weights, 
+                                                                                    best_obj, self.params['PBI_penalty'])
+                else:
+                    PBIS = np.fromiter(map(lambda solution: self.suboperators.apply['penalty_based_intersection'].apply(solution, weights[most_crowded_domain], 
+                                                                                                                        best_obj, self.params['PBI_penalty']),
+                                               last_level_by_domains[most_crowded_domain]), dtype = float)
+                    worst_solution = last_level_by_domains[most_crowded_domain][np.argmax(PBIS)]                    
+        
+        objective[1].delete_point(worst_solution)
+
+def use_item_if_no_default(arg : dict, key, replacement):
+    if key not in arg.keys():
+        arg[key] = replacement
+    return arg
+
+def get_basic_populator_updater(params : dict = {}, suboperators : dict = {}):
+    pop_updater = PopulationUpdater()
+    base_params = {'PBI_penalty' : 1}
+    pop_updater.params = params
+    
+    pop_updater.set_suboperators(operators = suboperators)
+
+
+
+class PopulationUpdaterConstrained(object):
+    def apply(self, objective, best_obj, weights):
+        '''
+        
+        Update population to get the pareto-nondomiated levels with the worst element removed. 
+        Here, "worst" means the solution with highest PBI value (penalty-based boundary intersection). 
+        Additionally, the constraint violations are considered in the selection of the 
+        "worst" individual.
+        
+        '''        
+        # objective = {offspring, pareto_levels}
+        
+        objective[1].update(objective[0])
+        cv_values = self.suboperators['constaint_violation'].apply(objective[0])
+        
+        if sum(cv_values) == 0:
+            if len(objective[1].levels) == 1:
+                worst_solution = self.suboperators['locate_pareto_worst'].apply(objective[1], weights, 
+                                                                                best_obj, self.params['PBI_penalty'])
+            else:
+                if objective[1].levels[len(objective[1].levels) - 1] == 1:
+                    domain_solutions = self.suboperators['population_to_sectors'].apply(objective[1].population, 
+                                                                                        weights)
+                    reference_solution = objective[1].levels[len(objective[1].levels) - 1][0]
+                    reference_solution_domain = [idx for idx in np.arange(domain_solutions) if reference_solution in domain_solutions[idx]]
+                    if len(domain_solutions[reference_solution_domain] == 1):
+                        worst_solution = self.suboperators['locate_pareto_worst'].apply(objective[1].levels, 
+                                                                                        weights, best_obj, 
+                                                                                        self.params['PBI_penalty'])
+                    else:
+                        worst_solution = reference_solution
+                else:
+                    last_level_by_domains = self.suboperators['population_to_sectors'].apply(objective[1].levels[len(objective[1].levels)-1], 
+                                                                                             self.weights)
+                    most_crowded_count = np.max([len(domain) for domain in last_level_by_domains]); 
+                    crowded_domains = [domain_idx for domain_idx in np.arange(len(self.weights)) 
+                                       if len(last_level_by_domains[domain_idx]) == most_crowded_count]
+    
+                    if len(crowded_domains) == 1:
+                        most_crowded_domain = crowded_domains[0]
+                    else:
+                        PBI = lambda domain_idx: np.sum([self.suboperators['penalty_based_intersection'].apply(sol_obj, self.weights[domain_idx], 
+                                                                                                               best_obj, self.params['PBI_penalty']) 
+                                                            for sol_obj in last_level_by_domains[domain_idx]])
+                        PBIS = np.fromiter(map(PBI, crowded_domains), dtype = float)
+                        most_crowded_domain = crowded_domains[np.argmax(PBIS)]
+                        
+                    if len(last_level_by_domains[most_crowded_domain]) == 1:
+                        worst_solution = self.suboperators['locate_pareto_worst'].apply(objective[1], weights, 
+                                                                                        best_obj, self.params['PBI_penalty'])
+                    else:
+                        PBIS = np.fromiter(map(lambda solution: self.suboperators['population_to_sectors'].apply(solution, self.weights[most_crowded_domain], 
+                                                                                                                 best_obj, self.params['PBI_penalty']), 
+                                               last_level_by_domains[most_crowded_domain]), dtype = float)
+                        worst_solution = last_level_by_domains[most_crowded_domain][np.argmax(PBIS)]                    
+        else:
+            infeasible = [solution for solution, _ in sorted(list(zip(objective[1].population, cv_values)), key = lambda pair: pair[1])]
+            infeasible.reverse()
+            infeasible = infeasible[:np.nonzero(cv_values)[0].size]
+            deleted = False
+            domain_solutions = self.suboperators['population_to_sectors'].apply(objective[1].population, weights)
+            
+            for infeasable_element in infeasible:
+                domain_idx = [domain_idx for domain_idx, domain in enumerate(domain_solutions) if infeasable_element in domain][0]
+                if len(domain_solutions[domain_idx]) > 1:
+                    deleted = True
+                    worst_solution = infeasable_element
+                    break
+            if not deleted:
+                worst_solution = infeasible[0]
+
+        objective[1].delete_point(worst_solution)
+
+
+class ConstraintViolationCalculator(object):
+    def __init__(self, constraints = None, param_keys = []):
+        assert all(isinstance(constraints, Constraint) for constr in constraints)
+        self._constraints = constraints
+        super().init(param_keys)
+    
+    def apply(self, pareto_levels, indexes = None):
+        def violation(individual):
+            return reduce(lambda y, z: y + z(individual.vals()), self._constraints, initial = 0)
+
+        if indexes is None:
+            indexes = np.arange(len(pareto_levels.population))
+        constraint_violations = np.array(list(map(lambda x: violation(pareto_levels.population[x]), indexes)))
+        
+        return constraint_violations
+    
+    def use_default_tags(self):
+        self._tags = {'constraint violation calculation', 'mixed input level', 'auxilary'}
